+++ python/ray/util/multiprocessing/__init__.py
@@ -1,0 +1,5 @@
+from multiprocessing import TimeoutError
+
+from .pool import Pool
+
+__all__ = ["Pool", "TimeoutError"]
